pyproject.toml
[project]
name = "nonebot-plugin-avatar-manager"
version = "0.1.0"
description = "NoneBot2 头像管理插件，支持机器人与群资料的立即修改和定时修改。"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "nonebot2>=2.3.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "pyright>=1.1.389",
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src"]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "pyright>=1.1.389",
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.pdm.scripts]
format = { composite = ["isort .", "black .", "ruff check . --fix"] }
lint = "ruff check ."
test = "pytest"

[tool.black]
line-length = 90
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "C", "T", "PYI", "Q"]
ignore = ["E402"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = []
plugin_dirs = ["src"]
builtin_plugins = []
